RG-4223:update

+++ src/main/java/in/wynk/payment/presentation/PaymentStatusPresentation.java
@@ -86,7 +86,7 @@ public class PaymentStatusPresentation implements IPaymentPresentation<AbstractC
         } else {
             PaidPack.PaidPackBuilder<?, ?> paidPackBuilder =
                     PaidPack.builder().title(offer.getTitle()).amount(plan.getFinalPrice()).period(plan.getPeriod().getValidity()).timeUnit(plan.getPeriod().getTimeUnit().name())
-                            .month(plan.getPeriod().getMonth()).perMonthValue(plan.getPrice().getMonthlyAmount()).dailyAmount(plan.getPrice().getDailyAmount()).day(plan.getPeriod().getDay())
+                            .month(plan.getPeriod().getMonth()).perMonthValue(plan.getPrice().getMonthlyAmount().intValue()).dailyAmount(plan.getPrice().getDailyAmount()).day(plan.getPeriod().getDay())
                             .currency(plan.getPrice().getCurrency()).isCombo(offer.isCombo());
             if (offer.isCombo()) {
                 final BundleBenefits.BundleBenefitsBuilder<?, ?> benefitsBuilder =
@@ -147,7 +147,7 @@ public class PaymentStatusPresentation implements IPaymentPresentation<AbstractC
         }
         return trialPackBuilder.paidPack(
                 PaidPack.builder().title(paidPlan.getTitle()).amount(paidPlan.getFinalPrice()).period(paidPlan.getPeriod().getValidity()).timeUnit(paidPlan.getPeriod().getTimeUnit().name())
-                        .month(paidPlan.getPeriod().getMonth()).perMonthValue(paidPlan.getPrice().getMonthlyAmount()).day(paidPlan.getPeriod().getDay())
+                        .month(paidPlan.getPeriod().getMonth()).perMonthValue(paidPlan.getPrice().getMonthlyAmount().intValue()).day(paidPlan.getPeriod().getDay())
                         .dailyAmount(paidPlan.getPrice().getDailyAmount()).currency(paidPlan.getPrice().getCurrency()).build()).isCombo(offer.isCombo()).build();
     }
